Michał Cichoń fixed integer clipping when drawing text.

+++ addons/font/allegro5/allegro_font.h
@@ -59,8 +59,8 @@ struct ALLEGRO_FONT_VTABLE
    ALLEGRO_FONT_METHOD(int, font_height, (const ALLEGRO_FONT *f));
    ALLEGRO_FONT_METHOD(int, char_length, (const ALLEGRO_FONT *f, int ch));
    ALLEGRO_FONT_METHOD(int, text_length, (const ALLEGRO_FONT *f, const ALLEGRO_USTR *text));
-   ALLEGRO_FONT_METHOD(int, render_char, (const ALLEGRO_FONT *f, int ch, int x, int y));
-   ALLEGRO_FONT_METHOD(int, render, (const ALLEGRO_FONT *f, const ALLEGRO_USTR *text, int x, int y));
+   ALLEGRO_FONT_METHOD(int, render_char, (const ALLEGRO_FONT *f, int ch, float x, float y));
+   ALLEGRO_FONT_METHOD(int, render, (const ALLEGRO_FONT *f, const ALLEGRO_USTR *text, float x, float y));
    ALLEGRO_FONT_METHOD(void, destroy, (ALLEGRO_FONT *f));
    ALLEGRO_FONT_METHOD(void, get_text_dimensions, (const ALLEGRO_FONT *f,
       const ALLEGRO_USTR *text, int *bbx, int *bby, int *bbw,
